requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
djangorestframework==3.12.2
djongo==1.3.4
easy-thumbnails==2.7.1
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
pymongo==3.11.3
pytz==2021.1
sqlparse==0.2.4
whitenoise==5.2.0
